+++ packages/tiptap-editor/web/src/components/RichContentAdapter/RichContentAdapter.tsx
@@ -254,5 +254,6 @@ export class RichContentAdapter implements TiptapAPI {
     setBlockType: () => {},
     setTextAlignment: () => {},
     _setSelection: () => {},
+    isAtomicBlockInSelection: () => false,
   };
 }