+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 
 - `toolbar-new`
   - [#3066](https://github.com/wix/ricos/pull/3066) fix color-picker modal style for mobile devices
+  - [#3075](https://github.com/wix/ricos/pull/3075) disable link when atomic block is selected
   - [#3078](https://github.com/wix/ricos/pull/3078) refactor toolbar modals position and width
 - `toolbars`
   - [#3063](https://github.com/wix/ricos/pull/3063) ui changes in static text toolbars (old/new toolbar)